Segfault in main, WordFind not created
In main when verifying that the board has been passed, we get a segfault. The current assumption is that it isn't actually passing the board.
WordFind doesn't work. it's still just a placeholder.
In order to work on WordFind we need to solve the issue in main, because wordFind requires both the board and trie.
Haven't tested trie in main yet.